Fix tests

+++ libsodium-sys/src/crypto_aead_chacha20poly1305.rs
@@ -4,7 +4,7 @@
 fn test_crypto_aead_chacha20poly1305_keybytes() {
     assert!(
         unsafe { crypto_aead_chacha20poly1305_keybytes() as usize }
-            == crypto_aead_chacha20poly1305_KEYBYTES
+            == crypto_aead_chacha20poly1305_KEYBYTES as usize
     )
 }
 
@@ -12,7 +12,7 @@ fn test_crypto_aead_chacha20poly1305_keybytes() {
 fn test_crypto_aead_chacha20poly1305_nsecbytes() {
     assert!(
         unsafe { crypto_aead_chacha20poly1305_nsecbytes() as usize }
-            == crypto_aead_chacha20poly1305_NSECBYTES
+            == crypto_aead_chacha20poly1305_NSECBYTES as usize
     )
 }
 
@@ -20,7 +20,7 @@ fn test_crypto_aead_chacha20poly1305_nsecbytes() {
 fn test_crypto_aead_chacha20poly1305_npubbytes() {
     assert!(
         unsafe { crypto_aead_chacha20poly1305_npubbytes() as usize }
-            == crypto_aead_chacha20poly1305_NPUBBYTES
+            == crypto_aead_chacha20poly1305_NPUBBYTES as usize
     )
 }
 
@@ -28,38 +28,38 @@ fn test_crypto_aead_chacha20poly1305_npubbytes() {
 fn test_crypto_aead_chacha20poly1305_abytes() {
     assert!(
         unsafe { crypto_aead_chacha20poly1305_abytes() as usize }
-            == crypto_aead_chacha20poly1305_ABYTES
+            == crypto_aead_chacha20poly1305_ABYTES as usize
     )
 }
 
 #[test]
 fn test_crypto_aead_chacha20poly1305_ietf_keybytes() {
     assert!(
-        unsafe { crypto_aead_chacha20poly1305_ietf_keybytes() as usize }
-            == crypto_aead_chacha20poly1305_ietf_KEYBYTES
+        unsafe { crypto_aead_chacha20poly1305_ietf_keybytes() }
+            == crypto_aead_chacha20poly1305_ietf_KEYBYTES as usize
     )
 }
 
 #[test]
 fn test_crypto_aead_chacha20poly1305_ietf_nsecbytes() {
     assert!(
-        unsafe { crypto_aead_chacha20poly1305_ietf_nsecbytes() as usize }
-            == crypto_aead_chacha20poly1305_ietf_NSECBYTES
+        unsafe { crypto_aead_chacha20poly1305_ietf_nsecbytes() }
+            == crypto_aead_chacha20poly1305_ietf_NSECBYTES as usize
     )
 }
 
 #[test]
 fn test_crypto_aead_chacha20poly1305_ietf_npubbytes() {
     assert!(
-        unsafe { crypto_aead_chacha20poly1305_ietf_npubbytes() as usize }
-            == crypto_aead_chacha20poly1305_ietf_NPUBBYTES
+        unsafe { crypto_aead_chacha20poly1305_ietf_npubbytes() }
+            == crypto_aead_chacha20poly1305_ietf_NPUBBYTES as usize
     )
 }
 
 #[test]
 fn test_crypto_aead_chacha20poly1305_ietf_abytes() {
     assert!(
-        unsafe { crypto_aead_chacha20poly1305_ietf_abytes() as usize }
-            == crypto_aead_chacha20poly1305_ietf_ABYTES
+        unsafe { crypto_aead_chacha20poly1305_ietf_abytes() }
+            == crypto_aead_chacha20poly1305_ietf_ABYTES as usize
     )
 }
