+++ libsodium-sys/src/crypto_auth_hmacsha256.rs
@@ -2,12 +2,13 @@
 
 #[test]
 fn test_crypto_auth_hmacsha256_bytes() {
-    assert!(unsafe { crypto_auth_hmacsha256_bytes() as usize } == crypto_auth_hmacsha256_BYTES)
+    assert!(unsafe { crypto_auth_hmacsha256_bytes() } == crypto_auth_hmacsha256_BYTES as usize)
 }
 
 #[test]
 fn test_crypto_auth_hmacsha256_keybytes() {
     assert!(
-        unsafe { crypto_auth_hmacsha256_keybytes() as usize } == crypto_auth_hmacsha256_KEYBYTES
+        unsafe { crypto_auth_hmacsha256_keybytes() as usize }
+            == crypto_auth_hmacsha256_KEYBYTES as usize
     )
 }
